xml/parsing/manakai/cdata-1.dat

#data
<![CDATA[]]>
#errors
1;1;cdata outside of root element;;;m
1;13;no root element;;;m
#document

#data
<![CDATA[  ]]>
#errors
1;1;cdata outside of root element;;;m
1;15;no root element;;;m
#document

#data
<![CDATA[ab]]>
#errors
1;1;cdata outside of root element;;;m
1;15;no root element;;;m
#document

#data
a<![CDATA[bb]]>ccc
#errors
1;1;text outside of root element;;;m
1;19;no root element;;;m
#document

#data
<p>a<![CDATA[bb]]>ccc</p>
#errors
#document
| <{} p>
|   "abbccc"

#data
<a/><![CDATA[]]>
#errors
1;5;cdata outside of root element;;;m
#document
| <{} a>

#data
<a/><![CDATA[  ]]>
#errors
1;5;cdata outside of root element;;;m
#document
| <{} a>

#data
<a/><![CDATA[aa]]>
#errors
1;5;cdata outside of root element;;;m
#document
| <{} a>

#data
<a/><![CDATA[]]>  a
#errors
1;5;cdata outside of root element;;;m
#document
| <{} a>

#data
<![CDATA[
#errors
1;1;cdata outside of root element;;;m
1;10;no mse;;;m
1;10;no root element;;;m
#document

#data
<p><![CDATA[
#errors
1;13;no mse;;;m
1;13;in body:#eof;;;m
#document
| <{} p>

#data
<![CDATA[a
#errors
1;11;no mse;;;m
1;1;cdata outside of root element;;;m
1;11;no root element;;;m
#document

#data
<p><![CDATA[a
#errors
1;14;no mse;;;m
1;14;in body:#eof;;;m
#document
| <{} p>
|   "a"

#data
<![CDATA[a]
#errors
1;12;no mse;;;m
1;1;cdata outside of root element;;;m
1;12;no root element;;;m
#document

#data
<p><![CDATA[a]
#errors
1;15;no mse;;;m
1;15;in body:#eof;;;m
#document
| <{} p>
|   "a]"

#data
<![CDATA[a]]
#errors
1;13;no mse;;;m
1;1;cdata outside of root element;;;m
1;13;no root element;;;m
#document

#data
<p><![CDATA[a]]
#errors
1;16;no mse;;;m
1;16;in body:#eof;;;m
#document
| <{} p>
|   "a]]"

#data
<![CDATA[a]]><![CDATA[b]]><a/>
#errors
1;1;cdata outside of root element;;;m
#document
| <{} a>

#data escaped
<![CDATA[\u0000]]>
#errors
1;1;cdata outside of root element;;;m
1;14;no root element;;;m
#document escaped

#data escaped
<!DOCTYPE p><![CDATA[\u0000]]><p></p>
#errors
1;13;cdata outside of root element;;;m
#document escaped
| <!DOCTYPE p>
| <{} p>

#data escaped
<p><![CDATA[\u0000]]></p>
#errors
1;4;NULL;;;m
#document escaped
| <{} p>
|   "\uFFFD"

#data escaped
<p></p><![CDATA[\u0000]]>
#errors
1;8;cdata outside of root element;;;m
#document escaped
| <{} p>

#data escaped
<p></p><![CDATA[a\u0000b]]>
#errors
1;8;cdata outside of root element;;;m
#document escaped
| <{} p>
